refactored channel message into own component
