+++ practice-for-week-19-python-project-skeleton/react-app/src/components/Channels/ChannelMessage.js
@@ -1,0 +1,62 @@
+import { useEffect, useState } from "react";
+import { useDispatch, useSelector } from "react-redux";
+import { useHistory, useParams } from "react-router-dom";
+import { useSocket } from "../../context/SocketContext";
+import {
+  deleteChannelMessage,
+  getChannelMessages,
+} from "../../store/channel_messages";
+
+import "./ChannelMessages.css";
+
+const ChannelMessage = ({ message }) => {
+  const dispatch = useDispatch();
+  const history = useHistory();
+  const user = useSelector((state) => state.session.user);
+  const { socket } = useSocket();
+
+  const handleDelete = async () => {
+    await dispatch(deleteChannelMessage(message?.id));
+    dispatch(getChannelMessages(message?.channelId));
+    const payload = {
+      channel_id: parseInt(message?.channelId),
+    };
+    socket.emit("channelmessage", payload);
+    return history.push(
+      `/servers/${message?.serverId}/channels/${message?.channelId}`
+    );
+  };
+  return (
+    <>
+      <div className="message">
+        <div className="inner-message">
+          {message.message_author.image_url ? (
+            <img
+              alt={message.id}
+              src={message.message_author.image_url}
+              className="author-message-image"
+            ></img>
+          ) : (
+            <div className="author-message-image default-image">
+              {message.message_author.username[0].toUpperCase()}
+            </div>
+          )}
+          <div className="message-text">
+            <p className="username-message">
+              {message.message_author.username}
+            </p>
+            <p className="message-body">{message.body}</p>
+            {/* </div> */}
+          </div>
+          <div className="flex-row-end trash">
+            {message?.user_id === user?.id && (
+              <i className="fa-regular fa-trash-can" onClick={handleDelete}></i>
+            )}
+          </div>
+        </div>
+      </div>
+    </>
+  );
+};
+
+export default ChannelMessage;
